// Relavant_Info/dist/api/relevantInfo/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Service
 * Service
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface RelevantInfoApi {
	postRelevantInfoCreate: (request: Api.RelevantInfoDto | undefined) => Promise<PostRelevantInfoCreateResponse>
	deleteRelevantInfoDelete: (id: string) => Promise<DeleteRelevantInfoDeleteResponse>
	getRelevantInfoGet: (id: string) => Promise<GetRelevantInfoGetResponse>
	getRelevantInfoGetAll: (limit: number | null | undefined, direction: Api.DirectionParamEnum | undefined, sortByField: string | null | undefined) => Promise<GetRelevantInfoGetAllResponse>
	putRelevantInfoUpdate: (request: Api.RelevantInfoDto | undefined) => Promise<PutRelevantInfoUpdateResponse>
}

export type PostRelevantInfoCreateResponse = PostRelevantInfoCreate201Response | PostRelevantInfoCreate401Response | PostRelevantInfoCreate404Response | PostRelevantInfoCreate422Response

export interface PostRelevantInfoCreate201Response {
	status: 201
	body: Api.RelevantInfoDto
	headers?: never
}

export interface PostRelevantInfoCreate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PostRelevantInfoCreate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PostRelevantInfoCreate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type DeleteRelevantInfoDeleteResponse = DeleteRelevantInfoDelete200Response | DeleteRelevantInfoDelete401Response | DeleteRelevantInfoDelete404Response | DeleteRelevantInfoDelete422Response

export interface DeleteRelevantInfoDelete200Response {
	status: 200
	body: Api.MessageDto
	headers?: never
}

export interface DeleteRelevantInfoDelete401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface DeleteRelevantInfoDelete404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface DeleteRelevantInfoDelete422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetRelevantInfoGetResponse = GetRelevantInfoGet200Response | GetRelevantInfoGet401Response | GetRelevantInfoGet404Response | GetRelevantInfoGet422Response

export interface GetRelevantInfoGet200Response {
	status: 200
	body: Api.RelevantInfoDto
	headers?: never
}

export interface GetRelevantInfoGet401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetRelevantInfoGet404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface GetRelevantInfoGet422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetRelevantInfoGetAllResponse = GetRelevantInfoGetAll200Response | GetRelevantInfoGetAll401Response | GetRelevantInfoGetAll404Response

export interface GetRelevantInfoGetAll200Response {
	status: 200
	body: Api.RelevantInfoPagedResultDto
	headers?: never
}

export interface GetRelevantInfoGetAll401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetRelevantInfoGetAll404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export type PutRelevantInfoUpdateResponse = PutRelevantInfoUpdate200Response | PutRelevantInfoUpdate401Response | PutRelevantInfoUpdate404Response | PutRelevantInfoUpdate422Response

export interface PutRelevantInfoUpdate200Response {
	status: 200
	body: Api.RelevantInfoDto
	headers?: never
}

export interface PutRelevantInfoUpdate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PutRelevantInfoUpdate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PutRelevantInfoUpdate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}